fakewallet: pay_if_regtest waits

+++ cashu/core/settings.py
@@ -132,8 +132,8 @@ class MintLimits(MintSettings):
 
 class FakeWalletSettings(MintSettings):
     fakewallet_brr: bool = Field(default=True)
-    fakewallet_delay_outgoing_payment: Optional[int] = Field(default=3)
-    fakewallet_delay_incoming_payment: Optional[int] = Field(default=3)
+    fakewallet_delay_outgoing_payment: Optional[float] = Field(default=3.0)
+    fakewallet_delay_incoming_payment: Optional[float] = Field(default=3.0)
     fakewallet_stochastic_invoice: bool = Field(default=False)
     fakewallet_payment_state: Optional[bool] = Field(default=None)
